requirements.txt
fastapi==0.115.11
uvicorn[standard]
pydantic
chromadb==1.0.9
torch
transformers
sentence_transformers
langchain_core
python-dotenv
openai
mcp